engine: Remove taa-no from Secure Skylake Server
The host with Secure Intel Icelake Server Family become non operational
because it does not provide "taa-no" CPU feature even though the
following command indicates that the host is not vulnerable.

cat /sys/devices/system/cpu/vulnerabilities/tsx_async_abort
Not affected

It is possible that this flag is not reported on systems that
are not vulnerable anymore. We disable TSX in our CPU definition
(we use Icelake-Server-noTSX) so it is not a security risk if we leave
the requirement for "taa-no" CPU feature.

Bug-Url: https://github.com/oVirt/ovirt-engine/issues/2184623
Signed-off-by: Lucia Jelinkova <ljelinko@redhat.com>