Allow Checkpoint removal to fail in some cases in DeleteAllVmCheckpointsCommand
If there is checkpoint inconsistency between the checkpoints in the
database and those on-disk, then we need to remove all checkpoints
on-disk and in the database.
But the on-disk removal can fail because the checkpoint is already gone
because of some corruption (for example due to an unclean shutdown).

Therefor we want to allow DeleteAllVmCheckpointsCommand to continue on
failure in some cases.

Signed-off-by: Jean-Louis Dupond <jean-louis@dupond.be>