core: Fail when getVolumeInfo fails during OVF read
In a scenario where we managed to list the images on the storage, but
failed to read get their information from vdsm, we will end up with
engine assuming the OVF_STOREs are missing.

Instead this patch treats a failure to retrieve the infromation as a
failure to force the user to retry.

Bug-Url: https://bugzilla.redhat.com/2244641
Signed-off-by: Benny Zlotnik <bzlotnik@redhat.com>