Added iSCSI multipath status to the API
In this commit the multipath status is available in the following
endpoint:
/ovirt-engine/api/hosts/xxxx/storage

A new field has been added <available_paths> that represents the
multipaths that are NOT faulty as an integer value. This is useful for monitoring (you can
compare 'available_paths' to 'paths' and trigger an alert when
'available_paths' is lower than 'paths'.

Signed-off-by: Brooklyn Dewolf <contact@brooklyn.gent>