bll/storage: don't check source domain space on lsm
When doing a LiveStorageMigration, the code checked if the disk/image
would fit not only on the destination, but also on the source.
This because theoretically the snapshot image can grow until the
disksize.

But this means that if we needed to live migrate a 1TB disk, we need
at least 1TB of free space on the source domain also.
This is quite useless, and causes problems to move disks from (for
example) an already overloaded source domain.

As this check is neither being done when creating a simple snapshot,
just drop the check on the source domain for LSM.

Signed-off-by: Jean-Louis Dupond <jean-louis@dupond.be>