packaging: enable engine when it was enabled in answer file
When using an answer file that contains:
QUESTION/1/OVESETUP_ENGINE_ENABLE=str:yes
The CoreEnv.ENABLE was not set to True if fqdn's didn't match and it was
not a new database.

But as we explicitly want to enable the engine according to the answer
file, we must enable it.

Signed-off-by: Jean-Louis Dupond <jean-louis@dupond.be>