core: reconnect faster to host after reboot
Altered sleepOnReboot to try and reach the host with a certain interval, ServerRebootSleepTime, as waiting time in between tries.
When the host is back online after reboot, cancel sleep timeout and continue with operations.
Instead of trying to reach the host after the ServerRebootTimeout has been reached, this is the max time the host has to come back online before breaking off operations.
Uses the stats it gets from the vdsproxy to check the state of the host and reports back that operations with this host can be continued.

Signed-off-by: Jasper Berton <jasper.berton@team.blue>