pom: align dependencies when using different maven versions
Set a parent in ovirt-findbugs-filters to make sure we pull same
dependencies there as in the root project.

Move ovirt-findbugs-filters to a plugin to avoid a cyclic dependency
when building.

Pin a version for maven-install-plugin.

This all to make sure we build using the same dependencies on all OS'es.
We need this for example on COPR to have correct dependencies included
in ovirt-engine-build-dependencies, as the srpm is build there on FC41
which has different plugin versions in its Super POM.

Signed-off-by: Jean-Louis Dupond <jean-louis@dupond.be>