core: added ability to alter video device through API
Return an additional video_type field in the display info for a VM, Template or Instance.
This next to the already existing graphic info and display type.

This field was added to have a difference between the QXL/VNC and VGA/VNC for example.
This exposes the chosen Video Type in the UI to the API side.

This also opens up the possibility to alter the Video Type for existing VM's for example from VGA/VNC to QXL/VNC.

Signed-off-by: Jasper Berton <jasper.berton@team.blue>