webadmin: added filter to display only not in use LUNs
An existing problem:
When creating a new storage domain, when selecting a LUN, both unused LUNs and used LUNs are shown in the list. Since the list of LUNs can be very large, it is very difficult to search for unused LUNs that can be selected for a storage domain.
This happens in the following cases:
	- selecting a LUN when creating a new storage domain;
	- selecting a LUN when editing an existing storage domain;
	- selecting a LUN when creating RDM disks for a VM.

Problem solving:
The filter "Hide used LUNs" is added to the list of LUNs. When this filter is checked, the list shows only unused LUNs that can be added to the storage domain. This filter is checked by default.

Signed-off-by: Evgeniy Kononov <ekononov1919@gmail.com>