core: deletion unused image links
Problem:
When creating a disk on a block storage domain (for example, FibreChannel), links to the disk are created on all active hosts that belong to the StoragePool of the disk being created. Further, when deleting a created disk, links to the disk are deleted only on the host with the SPM role, while incorrect links to the deleted disk remain on the other hosts.

Solution:
Added an API method on VDSM to remove unused links. When deleting a disk, the command to delete unused links is invoked for all active hosts that are associated with the StoragePool of the disk being deleted.

Signed-off-by: Evgeniy Kononov <ekononov1919@gmail.com>