core: allow trying to select new SPM when network problems with current SPM
For now there is some cases when SPM can be safely transited to another host when there is network problems with the current SPM.
But some checks not allow to start this process.
With this change we allow to start process to select new SPM.
If this process not safe, selection process will be stopped on next steps.

Signed-off-by: Stanislav Melnichuk <melnichuk.stas@gmail.com>