core: validation of volumeId in VmPayload
Volume Id is used as a volume label for the ISO (ISO9660).
The size of the label in ISO9660 is limited to a maximum of 32 characters and i18n type of valid characters, so this needs to be validated when this value is optionally set through the API.
Otherwise a VM becomes unbootable because mkisofs can't create an ISO from a given payload that strays from these restrictions.

Signed-off-by: Jasper Berton <jasper.berton@team.blue>