core: improve isLiveDelete check
Just checking if VMStatus is Up is not enough.

There are some cases where the VM is actually active, but the status is not Up.
For example when there is a lot of I/O and using thin provisioned disks on block
storage, the VM sometimes goes into paused state.
If you then remove a checkpoint, the code does a coldDeleteVmCheckpoint.
But this should never happen, as a cold delete uses qemu-img to delete the
checkpoint and if the qcow2 image is still in use by a qemu process this will
corrupt the qcow2 image.

Signed-off-by: Peter Boden <peter.boden@team.blue>