webadmin: fix template creation resource issues
Fixes an issue where creating a template from a VM or VM snapshot
can result in incorrect vCPU values and
potentially incorrect memory values in the template.

With this change, VmListModel.onNewTemplate() now runs model.validate()
so that "Create Template" could fail fast if the VM model is invalid,
instead of creating a template with wrong memory/CPU settings.

Furthermore, getTotalCpuCores() is now called inside a callback function
in order to avoid a race condition.

Signed-off-by: Guoyong Zhang <guoyong.zhang@oracle.com>