packaging: handle engine-setup deprecated features
The engine setup will warn that the GlusterFS and Cinderlib are unsupported
from RHEL 10.
The default app-mode is also changed from "Both" to "Virt".
Since GlusterFS has been removed it doesn't make sense to install new clusters with
it anymore. It is however still possible if you are e.g. installing on RHEL 9.

Signed-off-by: Peter Boden <peter.boden@team.blue>