core: add stringprep/scram to dependency
Adding stringprep/saslprep and scram(client/common) as dependencies in
the pom.
This is needed to fix local builds. Otherwise the jars for those modules
are missing.

Next to that also rename stringprep and saslprep with the stringprep
prefix as this is more clear they are in the same groupId.

Signed-off-by: Jean-Louis Dupond <jean-louis@dupond.be>