packaging: migrate legacy .p12 on openssl upgrade
When restoring a backup from CentOS 8 onto a CentOS 9 system, the p12 files that were generated are invalid,
because they use a legacy encryption on CentOS 9.

So, upon restore, we try to read the p12 files. When this fails we try to read the
p12 file with the `legacy` flag. If this succeeds, we re-encrypt the p12 file
with a newer encryption algorythm. The old p12 files is backup up.

We log all actions so that the user knows what happened, and can act upon errors.

Signed-off-by: Peter Boden <peter.boden@team.blue>