dev: quality of life improvements
Start the postgres container automatically when the devcontainer starts.
Also introduced the ability to have a separate postgresql version per
OS version.

Added `.devcontainer/compose-down.sh` script which should be called when
you want to switch to another OS version. This is needed to free up
certain ports on the host system (e.g. 5432 for postgres). Call the
script with the os version you want to bring down e.g. `centos10`.

Bake the PREFIX var into the container ans put it on the PATH by default.
Because of this you nog longer need to provide the PREFIX var in your commands.
So `make` can be run without the `PREFIX=...`, you can just run `engine-setup` without the full
path etc.

Signed-off-by: Peter Boden <peter.boden@team.blue>