dependencies: add deps for apache-commons-compress
Since apache-commons-compress 1.26.0, it depends on apache-commons-io /
codec and lang3.
As the version in CentOS 10 is 1.27.1 and on CentOS 9 its 1.21, we add
the dependencies as optional. This to make sure it just loads on CentOS
9 without the dependencies installed, while on CentOS 10 it will load
them. As those are dependencies on the package itself.

All 3 dependencies are also available on the Wildfly base libs, but
these are too old, which causes NoSuchClass exceptions when relying on
those. Therefore we use the versions shipped via rpm.

In the main pom.xml we align the versions to the versions in CentOS 10.

Signed-off-by: Jean-Louis Dupond <jean-louis@dupond.be>