setup: improve invalid postgres config message
If the `password_encryption` field is missing from the pg config,
also warn the user that pg_hba needs to be updated and role verifiers need
to be rotated.

When working on a local database the setup will ask if these settings should
be auto-updated, for a remote database the setup will exit.

Signed-off-by: Peter Boden <peter.boden@team.blue>