pom: build compress and deps on oldest runtime version
We need to build against the oldest runtime versions that we support.
So we need to use the versions used in RHEL9/CentOS9.

This because otherwise code can be incompatible with the library
version installed by the OS.

Change this for commons-compress and all it dependencies.
Those were bumped in c46cb556a69a1e728eff7b867888f783bd72a941

Signed-off-by: Jean-Louis Dupond <jean-louis@dupond.be>