packaging: add INSTALL_ROOT to engine-backup
Add the concept of an install root to the engine-backup.
On a regular production system the install root is `/`, but on
development systems this root (i.e. installed from an rpm) can be anything.
This makes it so that taking and restoring backups of development systems
is impossible.

There already was a devmode flag, if this flag is set to 1 then the
INSTALL_ROOT will be equal to `$PREFIX/`, in all other cases the root
will be `/` (like it was before).

Also, use the 'ENGINE_USR' var where a hardcoded path was used.

A backup that you take is now agnostic of the path where
ovirt was/is installed. So you can restore a production backup on a dev
box and vice versa.

Signed-off-by: Peter Boden <peter.boden@team.blue>