packaging: add shutdown hook
Add optional shutdown hook to pythonlib/service.py.
This can be used to implement a custom shutdown procedure for
an external process instead of sending a signal to it.
The signal will however still be used as a backup if something
goes wrong in said custom method.

The reason for adding this is signalhandler being removed from the
JDK, which makes proper signal processing harder.

Signed-off-by: Peter Boden <peter.boden@team.blue>