webadmin: remove ie10 from all-user-agents list
If BUILD_ALL_USER_AGENTS is set to 1 it builds with the all-user-agents list.
After the gwt version bump in https://github.com/oVirt/ovirt-engine/commit/db75424f2bcc the used gwt version went from 2.9.0 to 2.12.1.
In this newer version support was dropped to build for ie10, to ensure builds are still possible it needed to be removed from the list.

Signed-off-by: Jasper Berton <jasper.berton@team.blue>