qos: Update VM Network QoS on VM's when changed
Currently we don't trigger a refresh of all the running VM's when we
modify the QoS on a vNIC Profile.
Next to that, when adding or removing a QoS on a vNIC profile, an update
is not triggered neither.

We implement here that we update Vm Nic's in the following cases:
- When we modify the QoS used by the VM Nic
- When we add a QoS on the used vNIC Profile
- When we remove QoS on the used vNIC Profile
- When we use another QoS on the used vNIC Profile

As we now update the QoS live on a change, we don't need to mark the
vNIC as out of sync anymore on QoS update.

Signed-off-by: Jean-Louis Dupond <jean-louis@dupond.be>