pom: add versioned dependency on plexus-utils
On maven 3.6.3 (CentOs9), plexus-utils 1.1 is injected by default [1].
This conflicts with maven 3.9.9 on CentOS 10, as there it doesn't inject
plexus-utils 1.1. [2]

When ovirt-engine-build-dependencies srpm is build on CentOS 10, but
used on CentOS 9, the build fails because it's missing plexus-utils 1.1.

Fix this by adding a versioned plexus-utils depends on
maven-surefire-plugin. Then the correct version gets pulled in, and same
dependencies exist for centos9 or centos10.

[1]: https://github.com/apache/maven/blob/maven-3.6.3/maven-core/src/main/java/org/apache/maven/plugin/internal/PlexusUtilsInjector.java#L32-L37
[2]: https://github.com/apache/maven/pull/367

Signed-off-by: Jean-Louis Dupond <jean-louis@dupond.be>